requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-login==0.6.2
werkzeug==2.2.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
openai==0.27.8
requests==2.28.2
gunicorn==20.1.0
